% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd_to_decimal.R
\name{icd_to_decimal}
\alias{icd_to_decimal}
\title{Convert ICD-10 Codes to Standard Decimal Format}
\usage{
icd_to_decimal(icd_codes)
}
\arguments{
\item{icd_codes}{A character vector of ICD-10 codes.}
}
\value{
A character vector of ICD-10 codes in standard decimal format.
}
\description{
This function takes ICD-10 codes and converts them into the standard decimal format.
}
\examples{
icd_to_decimal(c("A010", "B20"))
}
